Rename helper method, move const, and de-dupe divider logic
