+++ public/app/features/browse-dashboards/state/actions.ts
@@ -1,4 +1,4 @@
-import { TEAM_FOLDERS_UID } from 'app/core/components/NestedFolderPicker/useTeamOwnedFolder';
+import { TEAM_FOLDERS_UID } from 'app/features/search/constants';
 import { GENERAL_FOLDER_UID } from 'app/features/search/constants';
 import { DashboardViewItem, DashboardViewItemKind } from 'app/features/search/types';
 import { createAsyncThunk } from 'app/types/store';
